better square up?

+++ src/main/java/frc/robot/subsystems/drive/DriveSubsystem.java
@@ -427,7 +427,7 @@ public class DriveSubsystem extends SubsystemBase {
 
 				double currentAngleRot = currentPose.getRotation().getRotations();
 
-				double targetAngleRot = Math.round(currentAngleRot * 4) / 4;
+				double targetAngleRot = ((double) Math.round(currentAngleRot * 4)) / 4;
 
 				double angularSpeedRps = rotationPID.calculate(currentAngleRot, targetAngleRot * 2 * Math.PI);
